+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to **Pipecat Cloud** will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+- Added an option to retrieve logs by session ID using `--session-id` / `-s`
+  with the `agent logs` command.
+
 ## [0.2.10] - 2025-10-25
 
 ### Removed